Fixes the problem with parsing just the first part of the document
And not the whole document. There's a new test inserted, new debugging
info all over the place, but it works only with the latest version of
Pod::Load, this has been bumped up in META6.json

This closes #55 (again). Feel free to open if it does not work for you.