Simplify `render` multi
There is no need for different candidates for Array and Pod::Block
if the implementation is the same. Create a single candidate for the
two types where `load` call is necessary and use nextwith to pass
the execution there. This makes code shorter and easier to think about,
while not losing a lot in readabilty.