Introduce TOC::Calculator
Prior this, the code gathered all the headers and rendered them into text in place.
While this is not really wrong, it is not future-proof, because actual content (TOC structure)
is tied to its presentation (particular HTML) and if the user wants to add a mere class or a wrapper
div, everything, including heading gathering logic, must be re-implemented, because do-toc sub
is not extendable.
So, here we introduce TOC::Calculator which has `calculate`, `render` and `render-heading` methods.
First one calculates a structure, representing the TOC of the $pod in question and two other
solely do rendering, thus making it easier to extend and digest. Also, increase test coverage a bit.