Improve stability in multi-threaded usage
Stress-testing shows `find-headings` method triggered numerous errors when run in
a multi-threaded env, so convert it into a method (this introduced additional API point, but as rewriting
it into `given/where` would be far less efficient and private multi are not supported, this is the way
we go here) and make the whole class a monitor that fixes some bugs.
Additionally, add a simple workaround for internal bugs triggered by Template::Mustache rendering in
concurrent environment.

This deals with the most annoying cases of concurrent (ab)use of Pod::To::HTML module, surviving
rendering whole documentation for 100 times with 24 parallel workers.