Pass config of a named paragraph to renderer
Pod paragraph in named blocks can be flexibly configured, but
a paragraph renderer won't see the config of its parent the way it
is not passed now.

We probably have to do this everywhere, but lazy evaluation of
such cases is likely a better idea to avoid clutter and rapid perf bumps.