Open a way to specify custom page renderers
Previously, a very hardcoded config with Template::Mustache in mind
was used. We preserve this as a backward-compatible default, but
introduce a new role `Pod::To::HTML::Renderer` which provides a simplest
working API for a custom page renderer.
This is still not completely flexible with regard to the rendering context passed,
but for now it allows downstream users to provide good enough custom page renderers.